refactor(cashflow-sankey): Phase 3 - Extract Custom Cashflow Sankey feature slice
- Created backend service backend/src/features/CashflowSankey/service.ts to extract layout math and munging logic.
- Defined shared types in shared/types/CashflowSankey.ts for node and link payloads.
- Exposed the service via backend/src/features/CashflowSankey/router.ts and registered it in the central registry.
- Wrote API client wrapper and React hook in frontend/src/features/CashflowSankey/api.ts.
- Created frontend/src/features/CashflowSankey/ui.tsx using SVG viewBox for responsive native scaling.
- Updated frontend/src/features/MonthlySpending/ui.tsx and frontend/src/App.tsx to use the new slice components.
- Removed legacy frontend/src/Sankey.tsx monolithic code.

Ref: Code is a mess, Embrace Vertical Slice #49
